new commit

+++ scripts/main.py
@@ -71,23 +71,22 @@ def main(arguments):
                                         f"\tPermission {permission_name} is set to {permission_mode} Successfully in {wb_id}\n")
                                 else:
                                     for permission in user_permissions:
-                                        if permission.get('name') == permission_name:
-                                            if permission.get('mode') != permission_mode:
-                                                existing_mode = permission.get(
-                                                    'mode')
-                                                delete_permission(
-                                                    data, auth_token, wb_id,
-                                                    permission_user_or_group_id, permission_name,
-                                                    existing_mode, version, is_group)
-                                                print(
-                                                    f"\tPermission {permission_name} : {existing_mode} is deleted Successfully in {wb_id}\n")
+                                        if permission.get('name') == permission_name and permission.get('mode') != permission_mode:
+                                            existing_mode = permission.get(
+                                                'mode')
+                                            delete_permission(
+                                                data, auth_token, wb_id,
+                                                permission_user_or_group_id, permission_name,
+                                                existing_mode, version, is_group)
+                                            print(
+                                                f"\tPermission {permission_name} : {existing_mode} is deleted Successfully in {wb_id}\n")
 
-                                                add_permission(
-                                                    data, wb_id, permission_user_or_group_id,
-                                                    version, auth_token, permission_name,
-                                                    permission_mode, is_group)
-                                                print(
-                                                    f"\tPermission {permission_name} is set to {permission_mode} Successfully in {wb_id}\n")
+                                            add_permission(
+                                                data, wb_id, permission_user_or_group_id,
+                                                version, auth_token, permission_name,
+                                                permission_mode, is_group)
+                                            print(
+                                                f"\tPermission {permission_name} is set to {permission_mode} Successfully in {wb_id}\n")
                         else:
                             logging.info(
                                 "Something went wrong, Error occured.\n User Name or List of Permissions in template are null")
